alternator: work around HTTP header case-sensitivity
It turns out that the official AWS SDK C++ driver sends
the X-Amz-Target header in lowercase, which triggers a bug
in Seastar HTTP server.
As a quick workaround to enable more drivers to work seamlessly
with alternator, all-lowercase and all-uppercase headers
are also checked just in case.

Tests: alternator(local)
Refs https://github.com/scylladb/seastar/issues/636
Fixes #6319