forbid tmpdir().path()
tmpdir is an object that exploits RAII for providing the user
with the temporary directory. however, it is very easy to miss
that it is must not be used as an rvalue for the path() call.

Implementation
==============

I wanted to forbid tmpdir to be an rvalue, but there are a lot
of places that use rvalue to lvalue materialization to future
with do_with(tmpdir(), [](){...).

The implementation deletes `path() &&` leaving only `path() &`.
It also extracts path generation logic to the separate static
method `tmpdir::temp_directory_path()`.