storage_proxy: local logic for two-phase truncate
Adds a truncate_session_manager object which keeps track of ongoing
truncations from a local point of view. It allows for performing both
phases of truncation:

- PREPARE: creates a session object for the truncation and sets a timer
  that removes it after timeout. If a session for this table exists, it
  is reused and timeout is reset. Generating and sending hints for the
  table and its views is blocked, hints are flushed and in-flight hints
  are waited for.
- ACCEPT: asserts that there is a live session object associated with
  the truncation of this table. Removes local table data and unblocks
  hint operations on the table and its views. Finally, the session
  object is also removed.

If truncation timeout is reached locally, hint operations are unblocked
and the session object is removed.

If one of the phases fail on some nodes, it might happen that truncate
sessions on some nodes will remain until their local expiry timer goes
off. In that case, the user can try again - PREPARE will re-use existing
sessions and will just reset local expiry timer.

A limitation of the design that allows re-using a session is that it
does not support concurrent truncations of the same table. For example,
some nodes may process ACCEPT from one coordinator earlier than the
other, and both coordinators will fail because they will receive session
expired errors from nodes that scheduled to process ACCEPT request
later.

This is not a problem because concurrent truncation of the same table,
while theoretically possible, does not make much sense, and the worst
that can happen is that hints will be blocked for a small period of
time. In such case, truncation should be retried, which will clean up
those sessions (if truncation succeeds).

Concurrent truncation of two different tables is perfectly possible and
should not interfere with each other.