cdc: remove the !bootstrap_tokens.empty() restriction
when creating a CDC generation. bootstrap_tokens is only required when
we create a new CDC generation on bootstrap, before joining the token
ring. But we would like to create new CDC generations after bootstrap
too (for rolling upgrade or on user requests). In these cases our tokens
are not special and can be obtained from token_metadata.